requirements.txt
torch>=2.1.0
torchvision>=0.16.0
albumentations>=1.3.1
opencv-python>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
matplotlib>=3.7.0
